README.md
# symbolic-programming-wmma-ysu
Mathematica (also known as Wolfram Mathematica) is a computational software program used in scientific, engineering, mathematical fields and other areas of technical computing. The Wolfram Language is the programming language used in Mathematica, which is a highly developed knowledge-based language that unifies a broad range of programming paradigms and uses its unique concept of symbolic programming to add a new level of flexibility to the very concept of programming. Symbolic programming is a programming paradigm in which the program can manipulate its own formulas and program components as if they were plain data. In this set of seminars, a general overview of the Wolfram Language will be discussed with focusing on symbolic programming.